added smoke test for OptionsBar
