+++ src/components/OptionsBar/OptionsBar.test.js
@@ -1,0 +1,11 @@
+import React from "react";
+import OptionsBar from './OptionsBar'
+import { shallow } from 'enzyme'
+
+describe('app component testing', function() {
+  it('renders successfully', function() {
+    const wrapper = shallow(<OptionsBar />); 
+
+    expect(wrapper.find('.option-bar').length).toEqual(1)
+  });
+});
