+++ src/components/OptionsBar/buttons/PlayButton/PlayButton.css
